refactor(webui): relocate system actions to TopBar
move reboot and refresh controls out of StatusTab and simplify layout
